+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `Orange Inverse` theme (Orange-OpenSource/ouds-ios#1038)
+- All color decorative tokens  (tokens library v1.7.0) (Orange-OpenSource/ouds-ios#1031)
 
 ## [0.19.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.18.0...0.19.0) - 2025-09-24